moved auto angle adjustment out into a command per DL
